refactor: validate, error, carBrand로 util 파일 분리
