+++ src/components/CardExpiryDate.tsx
@@ -3,7 +3,8 @@ import {
   MONTH_ERROR_MESSAGE,
   YEAR_ERROR_MESSAGE,
 } from '../constants/messages.ts';
-import { isMonthError, isYearError, isNumericInput } from '../utils/util.ts';
+import { isMonthError, isYearError } from '../utils/error.ts';
+import { isNumericInput } from '../utils/validate.ts';
 import type { CardExpiry, ExpireHandler } from '../types/cardStausTypes';
 import { useRef } from 'react';
