+++ src/utils/cardBrand.ts
@@ -1,7 +1,4 @@
-import type { DateError, MonthError, YearError } from '../types/errorTypes';
 import type { CardBrandType } from '../types/cardStausTypes';
-
-type ExpiryDateError = DateError | MonthError | YearError | 'normal';
 
 export function getCardBrand(cardNumber: string): CardBrandType {
   if (cardNumber.startsWith('4')) {
@@ -102,80 +99,4 @@ function isUnionPayPrefix(prefix: number): boolean {
     (prefix >= 624000 && prefix <= 626999) ||
     (prefix >= 628200 && prefix <= 628899)
   );
-}
-
-export function isMonthError(mode: DateError | MonthError | YearError | 'normal' | '') {
-  if (
-    mode === 'emptyBoth' ||
-    mode === 'emptyMonth' ||
-    mode === 'notMonthRange' ||
-    mode === 'notMonthNumber'
-  ) {
-    return true;
-  }
-  if (mode === 'normal' || mode === 'emptyYear' || mode === 'notYearNumber') {
-    return false;
-  }
-}
-
-export function isYearError(mode: DateError | MonthError | YearError | 'normal' | '') {
-  if (mode === 'emptyBoth' || mode === 'emptyYear' || mode === 'notYearNumber') {
-    return true;
-  }
-  if (
-    mode === 'normal' ||
-    mode === 'emptyMonth' ||
-    mode === 'notMonthRange' ||
-    mode === 'notMonthNumber'
-  ) {
-    return false;
-  }
-}
-
-export function isNumericInput(value: string): boolean {
-  return value === '' || /^\d+$/.test(value);
-}
-
-export function getExpiryDateChangeError(index: number, value: string): ExpiryDateError {
-  if (!isNumericInput(value)) {
-    return index === 0 ? 'notMonthNumber' : 'notYearNumber';
-  }
-
-  if (index === 0 && isInvalidMonth(value)) {
-    return 'notMonthRange';
-  }
-
-  return 'normal';
-}
-
-export function getMonthBlurError(month: string): ExpiryDateError {
-  if (isEmptyMonth(month)) {
-    return 'emptyMonth';
-  }
-
-  if (isInvalidMonth(month)) {
-    return 'notMonthRange';
-  }
-
-  return 'normal';
-}
-
-export function getYearBlurError(month: string, year: string): ExpiryDateError {
-  if (`${month}${year}`.length === 0) {
-    return 'emptyBoth';
-  }
-
-  if (year.length < 2) {
-    return 'emptyYear';
-  }
-
-  return getMonthBlurError(month);
-}
-
-function isEmptyMonth(month: string): boolean {
-  return month.length === 0 || month === '0' || month === '00';
-}
-
-function isInvalidMonth(month: string): boolean {
-  return Number(month) > 12 || month === '00';
 }
